Added BoxNum and changed GenerateProductTransaction to manually input box and feet

+++ src/main/java/Controllers/GenerateCustomerTransactController.java
@@ -657,7 +657,7 @@ public class GenerateCustomerTransactController {
                 dbExecuteTransaction.insertIntoDatabase(DBQueries.InsertQueries.Transaction.INSERT_INTO_TRANSACTION,
                         objects);
                 for(ProductTransaction tmp : transaction.getProductTransactionList()){
-                    int remain = tmp.getTotalNum() - tmp.getQuantity();
+                    float remain = tmp.getTotalNum() - tmp.getQuantity();
                     dbExecuteProduct.updateDatabase(DBQueries.UpdateQueries.Product.UPDATE_PRODUCT_QUANTITY,
                             remain, tmp.getProductId());
                 }
